Move shift click customization from Menu Entry Swapper to its own plugin and enable it by default
